requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
docutils==0.15
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
importlib-metadata==0.19
more-itertools==7.2.0
oauthlib==3.0.2
packaging==19.0
pbr==5.4.1
pluggy==0.12.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.1.1
pytest==5.0.1
pytest-cov==2.7.1
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
restructuredtext-lint==1.3.0
rsa==4.0
six==1.12.0
stevedore==1.30.1
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
